refactor : StoreRepository 파일 분리 (#323)
