+++ KCS/KCS/Domain/Interface/Repository/FetchSearchStoresRepository.swift
@@ -1,0 +1,19 @@
+//
+//  FetchSearchStoresRepository.swift
+//  KCS
+//
+//  Created by 조성민 on 1/11/24.
+//
+
+import RxSwift
+
+protocol FetchSearchStoresRepository {
+    
+    var storeStorage: StoreStorage { get }
+    
+    func fetchSearchStores(
+        location: Location,
+        keyword: String
+    ) -> Observable<[Store]>
+    
+}
